+++ django_geomultiplechoice/admin.py
