+++ django_geomultiplechoice/templates/geomultiplechoice_widget.html
