+++ django_geomultiplechoice/tests.py
