+++ django_geomultiplechoice/widgets.py
