+++ example/local_settings.example.py
@@ -13,7 +13,7 @@ ALLOWED_HOSTS = ['localhost']
 DATABASES = {
     'default': {
         'ENGINE': 'django.contrib.gis.db.backends.postgis',
-        'NAME': 'django-geomultiplechoicefield',
+        'NAME': 'django-geomultiplechoice',
         'USER': 'postgres',
         'PASSWORD': 'postgres',
         'HOST': 'postgres',
