+++ example/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for django_geomultiplechoicefield_project project.
+Django settings for example project.
 
 Generated by 'django-admin startproject' using Django 1.11.16.
 
@@ -26,7 +26,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'example',
-    'django_geomultiplechoicefield',
+    'django_geomultiplechoice',
     'leaflet',
 ]
 
@@ -40,7 +40,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'django_geomultiplechoicefield_project.urls'
+ROOT_URLCONF = 'example.urls'
 
 TEMPLATES = [
     {
@@ -58,7 +58,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'django_geomultiplechoicefield_project.wsgi.application'
+WSGI_APPLICATION = 'example.wsgi.application'
 
 # Password validation
 # https://docs.djangoproject.com/en/1.11/ref/settings/#auth-password-validators
