+++ example/templates/sample_form.html
@@ -32,7 +32,7 @@
     <form method="post" novalidate>
         {% csrf_token %}
         {{ form }}
-        <input type="submit" value="Submit" class="btn btn-success float-right">
+        <input type="submit" value="Submit" class="btn btn-primary float-right mt-3">
     </form>
   </div>
 </div>
