+++ example/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for django_geomultiplechoicefield_project project.
+WSGI config for example project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "django_geomultiplechoicefield_project.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "example.settings")
 
 application = get_wsgi_application()
